reorganize, start adding tests
