+++ scolapasta-unpack/src/lib.rs
@@ -1,769 +1,8 @@
-use core::num::IntErrorKind;
-use core::str;
-
-/// Enum representing different directives used for parsing format strings in
-/// Ruby's [`String#unpack`].
-///
-/// [`String#unpack`]: https://ruby-doc.org/core-3.1.2/String.html#method-i-unpack
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-pub enum Directive {
-    /// Integer directives
-    ///
-    /// Consists of various unsigned and signed integer directives.
-    ///
-    /// The corresponding characters for each directive are as follows:
-    ///
-    /// | Integer Directive | Returns  | Meaning                                                                      |
-    /// |-------------------|----------|------------------------------------------------------------------------------|
-    /// | `C`               | Integer  | 8-bit unsigned (`unsigned char`)                                             |
-    /// | `S`               | Integer  | 16-bit unsigned, native endian (`uint16_t`)                                  |
-    /// | `L`               | Integer  | 32-bit unsigned, native endian (`uint32_t`)                                  |
-    /// | `Q`               | Integer  | 64-bit unsigned, native endian (`uint64_t`)                                  |
-    /// | `J`               | Integer  | pointer width unsigned, native endian (`uintptr_t`)                          |
-    /// | `c`               | Integer  | 8-bit signed (`signed char`)                                                 |
-    /// | `s`               | Integer  | 16-bit signed, native endian (`int16_t`)                                     |
-    /// | `l`               | Integer  | 32-bit signed, native endian (`int32_t`)                                     |
-    /// | `q`               | Integer  | 64-bit signed, native endian (`int64_t`)                                     |
-    /// | `j`               | Integer  | pointer width signed, native endian (`intptr_t`)                             |
-    /// | `S_`, `S!`        | Integer  | `unsigned short`, native endian                                              |
-    /// | `I`, `I_`, `I!`   | Integer  | `unsigned int`, native endian                                                |
-    /// | `L_`, `L!`        | Integer  | `unsigned long`, native endian                                               |
-    /// | `Q_`, `Q!`        | Integer  | `unsigned long long`, native endian (`ArgumentError` if no `long long` type) |
-    /// | `J!`              | Integer  | `uintptr_t`, native endian (same as `J`)                                     |
-    /// | `s_`, `s!`        | Integer  | `signed short`, native endian                                                |
-    /// | `i`, `i_`, `i!`   | Integer  | `signed int`, native endian                                                  |
-    /// | `l_`, `l!`        | Integer  | `signed long`, native endian                                                 |
-    /// | `q_`, `q!`        | Integer  | `signed long long`, native endian (ArgumentError if no long long type)       |
-    /// | `j!`              | Integer  | `intptr_t`, native endian (same as `j`)                                      |
-    /// | `S>`, `s>`, `S!>`, `s!>`  | Integer  | same as directives without ">" except big endian. `S>` is the same as `n` |
-    /// | `L>`, `l>`, `L!>`, `l!>`  | Integer  | same as directives without ">" except big endian. `L>` is the same as `N` |
-    /// | `I!>`, `i!>`              | Integer  | same as directives without ">" except big endian                     |
-    /// | `Q>`, `q>`, `Q!>`, `q!>`  | Integer  | same as directives without ">" except big endian                     |
-    /// | `J>`, `j>`, `J!>`, `j!>`  | Integer  | same as directives without ">" except big endian                     |
-    /// | `S<`, `s<`, `S!<`, `s!<`  | Integer  | same as directives without "<" except little endian. `S<` is the same as `v` |
-    /// | `L<`, `l<`, `L!<`, `l!<`  | Integer  | same as directives without "<" except little endian. `L<` is the same as `V` |
-    /// | `I!<`, `i!<`              | Integer  | same as directives without "<" except little endian                  |
-    /// | `Q<`, `q<`, `Q!<`, `q!<`  | Integer  | same as directives without "<" except little endian                  |
-    /// | `J<`, `j<`, `J!<`, `j!<`  | Integer  | same as directives without "<" except little endian                  |
-    /// | `n`               | Integer  | 16-bit unsigned, network (big-endian) byte order                             |
-    /// | `N`               | Integer  | 32-bit unsigned, network (big-endian) byte order                             |
-    /// | `v`               | Integer  | 16-bit unsigned, VAX (little-endian) byte order                              |
-    /// | `V`               | Integer  | 32-bit unsigned, VAX (little-endian) byte order                              |
-    /// | `U`               | Integer  | UTF-8 character                                                              |
-    /// | `w`               | Integer  | BER-compressed integer (see [`Array#pack`])                                  |
-    ///
-    /// [`Array#pack`]: https://ruby-doc.org/core-3.1.2/Array.html#method-i-pack
-    Integer(IntegerDirective),
-
-    /// Float directives
-    ///
-    /// Consists of various float directives.
-    ///
-    /// The corresponding characters for each directive are as follows:
-    ///
-    /// | Float Directive | Returns  | Meaning                                                           |
-    /// |-----------------|----------|-------------------------------------------------------------------|
-    /// | `D`, `d`        | Float    | double-precision, native format                                   |
-    /// | `F`, `f`        | Float    | single-precision, native format                                   |
-    /// | `E`             | Float    | double-precision, little-endian byte order                        |
-    /// | `e`             | Float    | single-precision, little-endian byte order                        |
-    /// | `G`             | Float    | double-precision, network (big-endian) byte order                 |
-    /// | `g`             | Float    | single-precision, network (big-endian) byte order                 |
-    Float(FloatDirective),
-
-    /// String directives
-    ///
-    /// Consists of various string directives.
-    ///
-    /// The corresponding characters for each directive are as follows:
-    ///
-    /// | String Directive | Returns | Meaning                                                          |
-    /// |------------------|---------|------------------------------------------------------------------|
-    /// | `A`              | String  | arbitrary binary string (remove trailing nulls and ASCII spaces) |
-    /// | `a`              | String  | arbitrary binary string                                          |
-    /// | `Z`              | String  | null-terminated string                                           |
-    /// | `B`              | String  | bit string (MSB first)                                           |
-    /// | `b`              | String  | bit string (LSB first)                                           |
-    /// | `H`              | String  | hex string (high nibble first)                                   |
-    /// | `h`              | String  | hex string (low nibble first)                                    |
-    /// | `u`              | String  | UU-encoded string                                                |
-    /// | `M`              | String  | quoted-printable, MIME encoding (see RFC2045)                    |
-    /// | `m`              | String  | base64 encoded string (RFC 2045) (default)                       |
-    /// | `m0`             | String  | base64 encoded string (RFC 4648) if followed by 0                |
-    /// | `P`              | String  | pointer to a structure (fixed-length string)                     |
-    /// | `p`              | String  | pointer to a null-terminated string                              |
-    String(StringDirective),
-
-    /// Miscellaneous directives
-    ///
-    /// Consists of various miscellaneous directives.
-    ///
-    /// The corresponding characters for each directive are as follows:
-    ///
-    /// | Directive | Returns | Meaning                                         |
-    /// |-----------|---------|-------------------------------------------------|
-    /// | `@`       | ---     | skip to the offset given by the length argument |
-    /// | `X`       | ---     | skip backward one byte                          |
-    /// | `x`       | ---     | skip forward one byte                           |
-    Miscellaneous(MiscellaneousDirective),
-
-    /// Unknown directives
-    ///
-    /// Unpacking cannot fail. All unknown directives are ignored, but parsed as
-    /// valid directives.
-    ///
-    /// ```console
-    /// [3.2.2] > "aa".unpack('b-10b')
-    /// <internal:pack>:20: warning: unknown unpack directive '-' in 'b-10b'
-    /// => ["1", "1"]
-    /// ```
-    ///
-    /// All unknown directives print a warning:
-    ///
-    /// ```console
-    /// [3.2.2] > "aa".unpack('b-10{b')
-    /// <internal:pack>:20: warning: unknown unpack directive '-' in 'b-10{b'
-    /// <internal:pack>:20: warning: unknown unpack directive '{' in 'b-10{b'
-    /// => ["1", "1"]
-    /// ```
-    Unknown(u8),
-}
-
-impl Directive {
-    pub fn next_from_format_bytes(format: &mut &[u8]) -> Option<Self> {
-        let (&first, tail) = format.split_first()?;
-
-        let mut directive = Directive::from(first);
-
-        *format = tail;
-
-        if let Directive::Integer(ref mut directive) = directive {
-            directive.update_from_modifiers(format);
-        }
-        Some(directive)
-    }
-}
-
-impl From<u8> for Directive {
-    fn from(value: u8) -> Self {
-        if let Ok(directive) = IntegerDirective::try_from(value) {
-            return Self::Integer(directive);
-        }
-        if let Ok(directive) = FloatDirective::try_from(value) {
-            return Self::Float(directive);
-        }
-        if let Ok(directive) = StringDirective::try_from(value) {
-            return Self::String(directive);
-        }
-        if let Ok(directive) = MiscellaneousDirective::try_from(value) {
-            return Self::Miscellaneous(directive);
-        }
-        Self::Unknown(value)
-    }
-}
-
-/// Integer directives
-///
-/// Consists of various unsigned and signed integer directives.
-///
-/// The corresponding characters for each directive are as follows:
-///
-/// | Integer Directive | Returns  | Meaning                                                                      |
-/// |-------------------|----------|------------------------------------------------------------------------------|
-/// | `C`               | Integer  | 8-bit unsigned (`unsigned char`)                                             |
-/// | `S`               | Integer  | 16-bit unsigned, native endian (`uint16_t`)                                  |
-/// | `L`               | Integer  | 32-bit unsigned, native endian (`uint32_t`)                                  |
-/// | `Q`               | Integer  | 64-bit unsigned, native endian (`uint64_t`)                                  |
-/// | `J`               | Integer  | pointer width unsigned, native endian (`uintptr_t`)                          |
-/// | `c`               | Integer  | 8-bit signed (`signed char`)                                                 |
-/// | `s`               | Integer  | 16-bit signed, native endian (`int16_t`)                                     |
-/// | `l`               | Integer  | 32-bit signed, native endian (`int32_t`)                                     |
-/// | `q`               | Integer  | 64-bit signed, native endian (`int64_t`)                                     |
-/// | `j`               | Integer  | pointer width signed, native endian (`intptr_t`)                             |
-/// | `S_`, `S!`        | Integer  | `unsigned short`, native endian                                              |
-/// | `I`, `I_`, `I!`   | Integer  | `unsigned int`, native endian                                                |
-/// | `L_`, `L!`        | Integer  | `unsigned long`, native endian                                               |
-/// | `Q_`, `Q!`        | Integer  | `unsigned long long`, native endian (`ArgumentError` if no `long long` type) |
-/// | `J!`              | Integer  | `uintptr_t`, native endian (same as `J`)                                     |
-/// | `s_`, `s!`        | Integer  | `signed short`, native endian                                                |
-/// | `i`, `i_`, `i!`   | Integer  | `signed int`, native endian                                                  |
-/// | `l_`, `l!`        | Integer  | `signed long`, native endian                                                 |
-/// | `q_`, `q!`        | Integer  | `signed long long`, native endian (ArgumentError if no long long type)       |
-/// | `j!`              | Integer  | `intptr_t`, native endian (same as `j`)                                      |
-/// | `S>`, `s>`, `S!>`, `s!>`  | Integer  | same as directives without ">" except big endian. `S>` is the same as `n` |
-/// | `L>`, `l>`, `L!>`, `l!>`  | Integer  | same as directives without ">" except big endian. `L>` is the same as `N` |
-/// | `I!>`, `i!>`              | Integer  | same as directives without ">" except big endian                     |
-/// | `Q>`, `q>`, `Q!>`, `q!>`  | Integer  | same as directives without ">" except big endian                     |
-/// | `J>`, `j>`, `J!>`, `j!>`  | Integer  | same as directives without ">" except big endian                     |
-/// | `S<`, `s<`, `S!<`, `s!<`  | Integer  | same as directives without "<" except little endian. `S<` is the same as `v` |
-/// | `L<`, `l<`, `L!<`, `l!<`  | Integer  | same as directives without "<" except little endian. `L<` is the same as `V` |
-/// | `I!<`, `i!<`              | Integer  | same as directives without "<" except little endian                  |
-/// | `Q<`, `q<`, `Q!<`, `q!<`  | Integer  | same as directives without "<" except little endian                  |
-/// | `J<`, `j<`, `J!<`, `j!<`  | Integer  | same as directives without "<" except little endian                  |
-/// | `n`               | Integer  | 16-bit unsigned, network (big-endian) byte order                             |
-/// | `N`               | Integer  | 32-bit unsigned, network (big-endian) byte order                             |
-/// | `v`               | Integer  | 16-bit unsigned, VAX (little-endian) byte order                              |
-/// | `V`               | Integer  | 32-bit unsigned, VAX (little-endian) byte order                              |
-/// | `U`               | Integer  | UTF-8 character                                                              |
-/// | `w`               | Integer  | BER-compressed integer (see [`Array#pack`])                                  |
-///
-/// [`Array#pack`]: https://ruby-doc.org/core-3.1.2/Array.html#method-i-pack
-///
-/// These directives represent various integer types with different sizes, endianness,
-/// and additional miscellaneous types.
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-pub enum IntegerDirective {
-    /// 8-bit unsigned integer (`C`)
-    Unsigned8,
-
-    /// 16-bit unsigned, native endian (`S`)
-    Unsigned16NativeEndian,
-
-    /// 32-bit unsigned, native endian (`L`)
-    Unsigned32NativeEndian,
-
-    /// 64-bit unsigned, native endian (`Q`)
-    Unsigned64NativeEndian,
-
-    /// Pointer width unsigned, native endian (`J`)
-    ///
-    /// Also known as `J!`.
-    UnsignedPointerWidthNativeEndian,
-
-    /// 8-bit signed integer (`c`)
-    Signed8,
-
-    /// 16-bit signed, native endian (`s`)
-    Signed16NativeEndian,
-
-    /// 32-bit signed, native endian (`l`)
-    Signed32NativeEndian,
-
-    /// 64-bit signed, native endian (`q`)
-    Signed64NativeEndian,
-
-    /// Pointer width signed, native endian (`j`)
-    SignedPointerWidthNativeEndian,
-
-    /// 16-bit unsigned, big endian (`S>`)
-    Unsigned16BigEndian,
-
-    /// 32-bit unsigned, big endian (`L>`)
-    Unsigned32BigEndian,
-
-    /// 64-bit unsigned, big endian (`Q>`)
-    Unsigned64BigEndian,
-
-    /// Pointer width unsigned, big endian (`J>`)
-    UnsignedPointerWidthBigEndian,
-
-    /// 16-bit signed, big endian (`s>`)
-    Signed16BigEndian,
-
-    /// 32-bit signed, big endian (`l>`)
-    Signed32BigEndian,
-
-    /// 64-bit signed, big endian (`q>`)
-    Signed64BigEndian,
-
-    /// Pointer width signed, big endian (`j>`)
-    SignedPointerWidthBigEndian,
-
-    /// 16-bit unsigned, little endian (`S<`)
-    Unsigned16LittleEndian,
-
-    /// 32-bit unsigned, little endian (`L<`)
-    Unsigned32LittleEndian,
-
-    /// 64-bit unsigned, little endian (`Q<`)
-    Unsigned64LittleEndian,
-
-    /// Pointer width unsigned, little endian (`J<`)
-    UnsignedPointerWidthLittleEndian,
-
-    /// 16-bit signed, little endian (`s<`)
-    Signed16LittleEndian,
-
-    /// 32-bit signed, little endian (`l<`)
-    Signed32LittleEndian,
-
-    /// 64-bit signed, little endian (`q<`)
-    Signed64LittleEndian,
-
-    /// Pointer width signed, little endian (`j<`)
-    SignedPointerWidthLittleEndian,
-
-    /// `unsigned short`, native endian (`S_`, `S!`)
-    UnsignedShortNativeEndian,
-
-    /// `unsigned int`, native endian (`I`, `I_`, `I!`)
-    UnsignedIntNativeEndian,
-
-    /// `unsigned long`, native endian (`L_`, `L!`)
-    UnsignedLongNativeEndian,
-
-    /// `unsigned long long`, native endian (`Q_`, `Q!`)
-    ///
-    /// `ArgumentError` if the platform has no `long long` type.
-    UnsignedLongLongNativeEndian,
-
-    /// `signed short`, native endian (`s_`, `s!`)
-    SignedShortNativeEndian,
-
-    /// `signed int`, native endian (`i`, `i_`, `i!`)
-    SignedIntNativeEndian,
-
-    /// `signed long`, native endian (`l_`, `l!`)
-    SignedLongNativeEndian,
-
-    /// `signed long long`, native endian (`q_`, `q!`)
-    ///
-    /// `ArgumentError` if the platform has no `long long` type.
-    SignedLongLongNativeEndian,
-
-    /// `unsigned short`, big endian (`S!>`)
-    ///
-    /// Also known as `n`.
-    UnsignedShortBigEndian,
-
-    /// `unsigned int`, big endian (`I!>`)
-    UnsignedIntBigEndian,
-
-    /// `unsigned long`, big endian (`L!>`)
-    ///
-    /// Also known as `N`.
-    UnsignedLongBigEndian,
-
-    /// `unsigned long long`, big endian (`Q!>`)
-    ///
-    /// `ArgumentError` if the platform has no `long long` type.
-    UnsignedLongLongBigEndian,
-
-    /// `signed short`, big endian (`s!>`)
-    SignedShortBigEndian,
-
-    /// `signed int`, big endian (`i!>`)
-    SignedIntBigEndian,
-
-    /// `signed long`, big endian (`l!>`)
-    SignedLongBigEndian,
-
-    /// `signed long long`, big endian (`q!>`)
-    ///
-    /// `ArgumentError` if the platform has no `long long` type.
-    SignedLongLongBigEndian,
-
-    /// `unsigned short`, little endian (`S!<`)
-    ///
-    /// Also known as `v`.
-    UnsignedShortLittleEndian,
-
-    /// `unsigned int`, little endian (`I!<`)
-    UnsignedIntLittleEndian,
-
-    /// `unsigned long`, little endian (`L!<`)
-    ///
-    /// Also known as `V`.
-    UnsignedLongLittleEndian,
-
-    /// `unsigned long long`, little endian (`Q!<`)
-    ///
-    /// `ArgumentError` if the platform has no `long long` type.
-    UnsignedLongLongLittleEndian,
-
-    /// `signed short`, little endian (`s!<`)
-    SignedShortLittleEndian,
-
-    /// `signed int`, little endian (`i!<`)
-    SignedIntLittleEndian,
-
-    /// `signed long`, little endian (`l!<`)
-    SignedLongLittleEndian,
-
-    /// `signed long long`, little endian (`q!<`)
-    ///
-    /// `ArgumentError` if the platform has no `long long` type.
-    SignedLongLongLittleEndian,
-
-    /// 16-bit unsigned, network (big-endian) byte order (`n`)
-    ///
-    /// Also known as `S>`.
-    Unsigned16NetworkOrder,
-
-    /// 32-bit unsigned, network (big-endian) byte order (`N`)
-    ///
-    /// Also known as `L>`.
-    Unsigned32NetworkOrder,
-
-    /// 16-bit unsigned, VAX (little-endian) byte order (`v`)
-    ///
-    /// Also known as `S<`.
-    Unsigned16VaxOrder,
-
-    /// 32-bit unsigned, VAX (little-endian) byte order (`V`)
-    ///
-    /// Also known as `L<`.
-    Unsigned32VaxOrder,
-
-    /// UTF-8 character (`U`)
-    Utf8Character,
-
-    /// BER-compressed integer (`w`)
-    BerCompressedInteger,
-}
-
-impl TryFrom<u8> for IntegerDirective {
-    type Error = ();
-
-    fn try_from(value: u8) -> Result<Self, Self::Error> {
-        match value {
-            b'C' => Ok(Self::Unsigned8),
-            b'S' => Ok(Self::Unsigned16NativeEndian),
-            b'L' => Ok(Self::Unsigned32NativeEndian),
-            b'Q' => Ok(Self::Unsigned64NativeEndian),
-            b'J' => Ok(Self::UnsignedPointerWidthNativeEndian),
-            b'c' => Ok(Self::Signed8),
-            b's' => Ok(Self::Signed16NativeEndian),
-            b'l' => Ok(Self::Signed32NativeEndian),
-            b'q' => Ok(Self::Signed64NativeEndian),
-            b'j' => Ok(Self::SignedPointerWidthNativeEndian),
-            b'I' => Ok(Self::UnsignedIntNativeEndian),
-            b'i' => Ok(Self::SignedIntNativeEndian),
-            b'n' => Ok(Self::Unsigned16NetworkOrder),
-            b'N' => Ok(Self::Unsigned32NetworkOrder),
-            b'v' => Ok(Self::Unsigned16VaxOrder),
-            b'V' => Ok(Self::Unsigned32VaxOrder),
-            b'U' => Ok(Self::Utf8Character),
-            b'w' => Ok(Self::BerCompressedInteger),
-            _ => Err(()),
-        }
-    }
-}
-
-impl IntegerDirective {
-    #[allow(dead_code)]
-    fn next_from_format_bytes(bytes: &mut &[u8]) -> Option<Self> {
-        let (&first, tail) = bytes.split_first()?;
-
-        let mut directive = IntegerDirective::try_from(first).ok()?;
-
-        *bytes = tail;
-
-        directive.update_from_modifiers(bytes);
-        Some(directive)
-    }
-
-    fn update_from_modifiers(&mut self, format: &mut &[u8]) {
-        let mut chomp = 1_usize;
-        match (*self, *format) {
-            (_, []) => return,
-            (
-                Self::Unsigned16NativeEndian
-                | Self::UnsignedIntNativeEndian
-                | Self::Unsigned32NativeEndian
-                | Self::Unsigned64NativeEndian
-                | Self::UnsignedPointerWidthNativeEndian,
-                [b'!', b'>', ..],
-            ) => {
-                self.modify_platform_specific();
-                self.modify_big_endian();
-                chomp = 2;
-            }
-            (
-                Self::Unsigned16NativeEndian
-                | Self::UnsignedIntNativeEndian
-                | Self::Unsigned32NativeEndian
-                | Self::Unsigned64NativeEndian
-                | Self::UnsignedPointerWidthNativeEndian,
-                [b'!', b'<', ..],
-            ) => {
-                self.modify_platform_specific();
-                self.modify_little_endian();
-                chomp = 2;
-            }
-            (
-                Self::Unsigned16NativeEndian
-                | Self::UnsignedIntNativeEndian
-                | Self::Unsigned32NativeEndian
-                | Self::Unsigned64NativeEndian
-                | Self::Signed16NativeEndian
-                | Self::SignedIntNativeEndian
-                | Self::Signed32NativeEndian
-                | Self::Signed64NativeEndian,
-                [b'_' | b'!', ..],
-            ) => {
-                self.modify_platform_specific();
-            }
-            (Self::UnsignedPointerWidthNativeEndian | Self::SignedPointerWidthNativeEndian, [b'!', ..]) => {
-                self.modify_platform_specific();
-            }
-            (
-                Self::Unsigned16NativeEndian
-                | Self::UnsignedIntNativeEndian
-                | Self::Unsigned32NativeEndian
-                | Self::Unsigned64NativeEndian
-                | Self::UnsignedPointerWidthNativeEndian,
-                [b'>', ..],
-            ) => {
-                self.modify_big_endian();
-            }
-            (
-                Self::Unsigned16NativeEndian
-                | Self::UnsignedIntNativeEndian
-                | Self::Unsigned32NativeEndian
-                | Self::Unsigned64NativeEndian
-                | Self::UnsignedPointerWidthNativeEndian,
-                [b'<', ..],
-            ) => {
-                self.modify_little_endian();
-            }
-
-            // Consume unknown modifiers, emit a warning, and continue:
-            //
-            // ```console
-            // [3.2.2] > "1111111111111111".unpack('s-')
-            // <internal:pack>:20: warning: unknown unpack directive '-' in 's-'
-            // ```
-            _ch => {
-                // TODO: emit warning with `ch`.
-            }
-        }
-        *format = &format[chomp..];
-    }
-
-    fn modify_little_endian(&mut self) {
-        match self {
-            Self::Unsigned16NativeEndian => *self = Self::Unsigned16LittleEndian,
-            Self::Unsigned32NativeEndian => *self = Self::Unsigned32LittleEndian,
-            Self::Unsigned64NativeEndian => *self = Self::Unsigned64LittleEndian,
-            Self::Signed16NativeEndian => *self = Self::Signed16LittleEndian,
-            Self::Signed32NativeEndian => *self = Self::Signed32LittleEndian,
-            Self::Signed64NativeEndian => *self = Self::Signed64LittleEndian,
-            Self::UnsignedShortNativeEndian => *self = Self::UnsignedShortLittleEndian,
-            Self::UnsignedIntNativeEndian => *self = Self::UnsignedIntLittleEndian,
-            Self::UnsignedLongNativeEndian => *self = Self::UnsignedLongLittleEndian,
-            Self::UnsignedLongLongNativeEndian => *self = Self::UnsignedLongLongLittleEndian,
-            Self::SignedShortNativeEndian => *self = Self::SignedShortLittleEndian,
-            Self::SignedIntNativeEndian => *self = Self::SignedIntLittleEndian,
-            Self::SignedLongNativeEndian => *self = Self::SignedLongLittleEndian,
-            Self::SignedLongLongNativeEndian => *self = Self::SignedLongLongLittleEndian,
-            _ => {} // No modification needed for other variants
-        }
-    }
-
-    fn modify_big_endian(&mut self) {
-        match self {
-            Self::Unsigned16NativeEndian => *self = Self::Unsigned16BigEndian,
-            Self::Unsigned32NativeEndian => *self = Self::Unsigned32BigEndian,
-            Self::Unsigned64NativeEndian => *self = Self::Unsigned64BigEndian,
-            Self::Signed16NativeEndian => *self = Self::Signed16BigEndian,
-            Self::Signed32NativeEndian => *self = Self::Signed32BigEndian,
-            Self::Signed64NativeEndian => *self = Self::Signed64BigEndian,
-            Self::UnsignedShortNativeEndian => *self = Self::UnsignedShortBigEndian,
-            Self::UnsignedIntNativeEndian => *self = Self::UnsignedIntBigEndian,
-            Self::UnsignedLongNativeEndian => *self = Self::UnsignedLongBigEndian,
-            Self::UnsignedLongLongNativeEndian => *self = Self::UnsignedLongLongBigEndian,
-            Self::SignedShortNativeEndian => *self = Self::SignedShortBigEndian,
-            Self::SignedIntNativeEndian => *self = Self::SignedIntBigEndian,
-            Self::SignedLongNativeEndian => *self = Self::SignedLongBigEndian,
-            Self::SignedLongLongNativeEndian => *self = Self::SignedLongLongBigEndian,
-            _ => {} // No modification needed for other variants
-        }
-    }
-
-    fn modify_platform_specific(&mut self) {
-        match self {
-            Self::Unsigned16NativeEndian => *self = Self::UnsignedShortNativeEndian,
-            Self::Unsigned32NativeEndian => *self = Self::UnsignedLongNativeEndian,
-            Self::Unsigned64NativeEndian => *self = Self::UnsignedLongLongNativeEndian,
-            Self::Signed16NativeEndian => *self = Self::SignedShortNativeEndian,
-            Self::Signed32NativeEndian => *self = Self::SignedLongNativeEndian,
-            Self::Signed64NativeEndian => *self = Self::SignedLongLongNativeEndian,
-            Self::Unsigned16BigEndian => *self = Self::UnsignedShortBigEndian,
-            Self::Unsigned32BigEndian => *self = Self::UnsignedLongBigEndian,
-            Self::Unsigned64BigEndian => *self = Self::UnsignedLongLongBigEndian,
-            Self::Signed16BigEndian => *self = Self::SignedShortBigEndian,
-            Self::Signed32BigEndian => *self = Self::SignedLongBigEndian,
-            Self::Signed64BigEndian => *self = Self::SignedLongLongBigEndian,
-            Self::Unsigned16LittleEndian => *self = Self::UnsignedShortLittleEndian,
-            Self::Unsigned32LittleEndian => *self = Self::UnsignedLongLittleEndian,
-            Self::Unsigned64LittleEndian => *self = Self::UnsignedLongLongLittleEndian,
-            Self::Signed16LittleEndian => *self = Self::SignedShortLittleEndian,
-            Self::Signed32LittleEndian => *self = Self::SignedLongLittleEndian,
-            Self::Signed64LittleEndian => *self = Self::SignedLongLongLittleEndian,
-            _ => {} // No modification needed for other variants
-        }
-    }
-}
-
-/// Float directives
-///
-/// Consists of various float directives.
-///
-/// The corresponding characters for each directive are as follows:
-///
-/// | Float Directive | Returns  | Meaning                                                           |
-/// |-----------------|----------|-------------------------------------------------------------------|
-/// | `D`, `d`        | Float    | double-precision, native format                                   |
-/// | `F`, `f`        | Float    | single-precision, native format                                   |
-/// | `E`             | Float    | double-precision, little-endian byte order                        |
-/// | `e`             | Float    | single-precision, little-endian byte order                        |
-/// | `G`             | Float    | double-precision, network (big-endian) byte order                 |
-/// | `g`             | Float    | single-precision, network (big-endian) byte order                 |
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-pub enum FloatDirective {
-    /// Double-precision, native format (`D`, `d`)
-    DoubleNativeEndian,
-
-    /// Single-precision, native format (`F`, `f`)
-    SingleNativeEndian,
-
-    /// Double-precision, little-endian byte order (`E`)
-    DoubleLittleEndian,
-
-    /// Single-precision, little-endian byte order (`e`)
-    SingleLittleEndian,
-
-    /// Double-precision, network (big-endian) byte order (`G`)
-    DoubleBigEndian,
-
-    /// Single-precision, network (big-endian) byte order (`g`)
-    SingleBigEndian,
-}
-
-impl TryFrom<u8> for FloatDirective {
-    type Error = ();
-
-    fn try_from(value: u8) -> Result<Self, Self::Error> {
-        match value {
-            b'D' | b'd' => Ok(Self::DoubleNativeEndian),
-            b'F' | b'f' => Ok(Self::SingleNativeEndian),
-            b'E' => Ok(Self::DoubleLittleEndian),
-            b'e' => Ok(Self::SingleLittleEndian),
-            b'G' => Ok(Self::DoubleBigEndian),
-            b'g' => Ok(Self::SingleBigEndian),
-            _ => Err(()),
-        }
-    }
-}
-
-/// String directives
-///
-/// Consists of various string directives.
-///
-/// The corresponding characters for each directive are as follows:
-///
-/// | String Directive | Returns | Meaning                                                          |
-/// |------------------|---------|------------------------------------------------------------------|
-/// | `A`              | String  | arbitrary binary string (remove trailing nulls and ASCII spaces) |
-/// | `a`              | String  | arbitrary binary string                                          |
-/// | `Z`              | String  | null-terminated string                                           |
-/// | `B`              | String  | bit string (MSB first)                                           |
-/// | `b`              | String  | bit string (LSB first)                                           |
-/// | `H`              | String  | hex string (high nibble first)                                   |
-/// | `h`              | String  | hex string (low nibble first)                                    |
-/// | `u`              | String  | UU-encoded string                                                |
-/// | `M`              | String  | quoted-printable, MIME encoding (see RFC2045)                    |
-/// | `m`              | String  | base64 encoded string (RFC 2045) (default)                       |
-/// | `m0`             | String  | base64 encoded string (RFC 4648) if followed by 0                |
-/// | `P`              | String  | pointer to a structure (fixed-length string)                     |
-/// | `p`              | String  | pointer to a null-terminated string                              |
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-pub enum StringDirective {
-    /// Arbitrary binary string, remove trailing nulls and ASCII spaces (`A`)
-    ArbitraryBinaryTrimmed,
-
-    /// Arbitrary binary string (`a`)
-    ArbitraryBinary,
-
-    /// Null-terminated string (`Z`)
-    NullTerminated,
-
-    /// Bit string, MSB first (`B`)
-    BitStringMsbFirst,
-
-    /// Bit string, LSB first (`b`)
-    BitStringLsbFirst,
-
-    /// Hex string, high nibble first (`H`)
-    HexStringHighNibbleFirst,
-
-    /// Hex string, low nibble first (`h`)
-    HexStringLowNibbleFirst,
-
-    /// UU-encoded string (`U`)
-    UuEncoded,
-
-    /// Quoted-printable, MIME encoding - see RFC2045 (`M`)
-    QuotedPrintable,
-
-    /// Base64 encoded string - RFC 2045 default, RFC 4648 if followed by 0 (`m`)
-    Base64Encoded,
-
-    /// Pointer to a structure, fixed-length string (`P`)
-    StructurePointer,
-
-    /// Pointer to a null-terminated string (`p`)
-    NullTerminatedStringPointer,
-}
-
-impl TryFrom<u8> for StringDirective {
-    type Error = ();
-
-    fn try_from(value: u8) -> Result<Self, Self::Error> {
-        match value {
-            b'A' => Ok(Self::ArbitraryBinaryTrimmed),
-            b'a' => Ok(Self::ArbitraryBinary),
-            b'Z' => Ok(Self::NullTerminated),
-            b'B' => Ok(Self::BitStringMsbFirst),
-            b'b' => Ok(Self::BitStringLsbFirst),
-            b'H' => Ok(Self::HexStringHighNibbleFirst),
-            b'h' => Ok(Self::HexStringLowNibbleFirst),
-            b'u' => Ok(Self::UuEncoded),
-            b'M' => Ok(Self::QuotedPrintable),
-            b'm' => Ok(Self::Base64Encoded),
-            b'P' => Ok(Self::StructurePointer),
-            b'p' => Ok(Self::NullTerminatedStringPointer),
-            _ => Err(()),
-        }
-    }
-}
-
-/// Miscellaneous directives
-///
-/// Consists of various miscellaneous directives.
-///
-/// The corresponding characters for each directive are as follows:
-///
-/// | Directive | Returns | Meaning                                         |
-/// |-----------|---------|-------------------------------------------------|
-/// | `@`       | ---     | skip to the offset given by the length argument |
-/// | `X`       | ---     | skip backward one byte                          |
-/// | `x`       | ---     | skip forward one byte                           |
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-pub enum MiscellaneousDirective {
-    /// Skip to offset (`@`)
-    SkipToOffset,
-
-    /// Skip backward (`X`)
-    SkipBackward,
-
-    /// Skip forward (`x`)
-    SkipForward,
-}
-
-impl TryFrom<u8> for MiscellaneousDirective {
-    type Error = ();
-
-    fn try_from(value: u8) -> Result<Self, Self::Error> {
-        match value {
-            b'@' => Ok(Self::SkipToOffset),
-            b'X' => Ok(Self::SkipBackward),
-            b'x' => Ok(Self::SkipForward),
-            _ => Err(()),
-        }
-    }
-}
+mod directive;
+mod repetition;
+
+use directive::{Directive, MiscDirective};
+use repetition::Repetition;
 
 pub struct RangeError {
     message: &'static str,
@@ -779,114 +18,11 @@ impl RangeError {
     }
 }
 
-/// Represents various  directives used in a format string.
-///
-/// The unpack amount must be the last modifier:
-///
-/// ```console
-/// [3.2.2] > "1111111111111111".unpack('s*_')
-/// <internal:pack>:20: warning: unknown unpack directive '_' in 's*_'
-/// => [12593, 12593, 12593, 12593, 12593, 12593, 12593, 12593]
-/// [3.2.2] > "1111111111111111".unpack('s_*')
-/// => [12593, 12593, 12593, 12593, 12593, 12593, 12593, 12593]
-/// ```
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-pub enum UnpackAmount {
-    /// Specifies the number of times to repeat the preceding directive.
-    Repeat(usize),
-
-    /// Consume to end.
-    ///
-    /// An asterisk (`*`) will use up all remaining elements.
-    ConsumeToEnd,
-
-    /// The amount to unpack is consumed.
-    Finished,
-}
-
-impl Default for UnpackAmount {
-    fn default() -> Self {
-        Self::Repeat(1)
-    }
-}
-
-impl UnpackAmount {
-    pub fn from_format_bytes(format: &mut &[u8]) -> Result<Self, RangeError> {
-        match *format {
-            [] => return Ok(Self::default()),
-            [b'*', tail @ ..] => {
-                *format = tail;
-                return Ok(Self::ConsumeToEnd);
-            }
-            [byte, ..] if !byte.is_ascii_digit() => return Ok(Self::default()),
-            _ => {}
-        }
-        // Trim leading zeros:
-        //
-        // ```console
-        // [3.2.2] > "11111111111111111111111111111111".unpack('h010')
-        // => ["1313131313"]
-        // [3.2.2] > "11111111111111111111111111111111".unpack('h010').first.length
-        // => 10
-        // ```
-        while let Some((&b'0', tail)) = format.split_first() {
-            *format = tail;
-        }
-
-        let end_pos = format
-            .iter()
-            .position(|b| !b.is_ascii_digit())
-            .unwrap_or_else(|| format.len());
-        let number = str::from_utf8(&format[..end_pos]).expect("slice is only ASCII digits");
-
-        *format = &format[end_pos..];
-
-        match usize::from_str_radix(number, 10) {
-            Ok(count) => match NonZeroUsize::new(count) {
-                Some(count) => Ok(Self::Repeat(count)),
-                None => Ok(Self::Finished),
-            },
-            // we had some numbers, but they must have all been zero and stripped above.
-            Err(err) if matches!(err.kind(), IntErrorKind::Empty | IntErrorKind::Zero) => Ok(Self::Finished),
-            Err(err) if *err.kind() == IntErrorKind::PosOverflow => {
-                // ```
-                // <internal:pack>:20:in `unpack': pack length too big (RangeError)
-                // ```
-                Err(RangeError::with_message("pack length too big"))
-            }
-            Err(err) if *err.kind() == IntErrorKind::InvalidDigit => {
-                unreachable!("unexpected digits. Got {number}, expected 0-9 digits");
-            }
-            Err(err) if *err.kind() == IntErrorKind::NegOverflow => {
-                unreachable!("parsing into usize does not permit negative values");
-            }
-            // handle non-exhaustive case for this enum.
-            Err(_) => Ok(Self::Finished),
-        }
-    }
-
-    fn consume(&mut self) {
-        match self {
-            Self::Repeat(count) => {
-                let count = count.get();
-                match count.checked_sub(1) {
-                    None | Some(0) => *self = Self::Finished,
-                    Some(updated_count) => {
-                        *self = Self::Repeat(NonZeroUsize::new(updated_count).expect("count is nonzero"))
-                    }
-                }
-            }
-            // No updates.
-            Self::ConsumeToEnd | Self::Finished => {}
-        }
-    }
-}
-
 /// Iterator that parses a format string into directives.
 pub struct FormatStringIterator<'a> {
     bytes: &'a [u8],
     idx: usize,
-    last_directive: Option<(Directive, UnpackAmount)>,
+    last_directive: Option<(Directive, Repetition)>,
 }
 
 impl<'a> FormatStringIterator<'a> {
@@ -903,54 +39,43 @@ impl<'a> Iterator for FormatStringIterator<'a> {
     type Item = Result<Directive, RangeError>;
 
     fn next(&mut self) -> Option<Self::Item> {
-        if let Some((directive, mut unpack_amount)) = self.last_directive {
-            unpack_amount.consume();
-            if let UnpackAmount::Finished = unpack_amount {
-                self.last_directive = None;
+        if let Some((directive, repetition)) = self.last_directive.take() {
+            match repetition.next() {
+                None | Some(Repetition::Finished) => {}
+                Some(next_repetition) => {
+                    self.last_directive = Some((directive, next_repetition));
+                }
             }
             return Some(Ok(directive));
         }
 
         let misc = None;
         loop {
-            pub enum MiscellaneousDirective {
-                /// Skip to offset (`@`)
-                SkipToOffset,
-
-                /// Skip backward (`X`)
-                SkipBackward,
-
-                /// Skip forward (`x`)
-                SkipForward,
-            }
             let s = match misc {
-                Some((_, UnpackAmount::ConsumeToEnd)) => {
+                Some((_, Repetition::ConsumeToEnd)) => {
                     self.idx = usize::MAX;
                     &[]
                 }
-                Some((MiscellaneousSpecifier::SkipToOffset, UnpackAmount::Finished)) => {
+                Some((MiscDirective::SkipToOffset, Repetition::Finished)) => {
                     self.idx = 0;
                     self.bytes
                 }
-                Some((MiscellaneousSpecifier::SkipToOffset, UnpackAmount::Repeat(idx))) => {
+                Some((MiscDirective::SkipToOffset, Repetition::Repeat(idx))) => {
                     self.idx = idx.get();
                     match self.bytes.get(self.idx..) {
                         Some(s) => s,
                         None => return Some(Err(RangeError::with_message("@ outside of string"))),
                     }
                 }
-                Some((
-                    MiscellaneousSpecifier::SkipBackward | MiscellaneousSpecifier::SkipForward,
-                    UnpackAmount::Finished,
-                )) => self.bytes,
-                Some((MiscellaneousSpecifier::SkipBackward, UnpackAmount::Repeat(idx))) => {
+                Some((MiscDirective::SkipBackward | MiscDirective::SkipForward, Repetition::Finished)) => self.bytes,
+                Some((MiscDirective::SkipBackward, Repetition::Repeat(idx))) => {
                     self.idx = idx.get();
                     match self.bytes.get(self.idx..) {
                         Some(s) => s,
                         None => return Some(Err(RangeError::with_message("@ outside of string"))),
                     }
                 }
-                Some((MiscellaneousSpecifier::SkipForward, UnpackAmount::Repeat(idx))) => {
+                Some((MiscDirective::SkipForward, Repetition::Repeat(idx))) => {
                     self.idx = idx.get();
                     match self.bytes.get(self.idx..) {
                         Some(s) => s,
@@ -962,7 +87,7 @@ impl<'a> Iterator for FormatStringIterator<'a> {
 
             let mut format = s;
             let directive = Directive::next_from_format_bytes(&mut format);
-            let amount = match UnpackAmount::next_from_format_bytes(&mut format) {
+            let amount = match Repetition::next_from_format_bytes(&mut format) {
                 Ok(amount) => amount,
                 Err(err) => {
                     self.idx = usize::MAX;
